microservices/ai-service/requirements.txt
# microservices/ai-service/requirements.txt
Django==5.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
celery==5.3.4
django-celery-results==2.5.1
psycopg2-binary==2.9.10
redis==5.0.1
requests==2.31.0
python-dotenv==1.1.1
setuptools>=68.0.0
pytest==7.4.3
pytest-django==4.7.0
